Test Facility class
